requirements.txt
numpy
pandas
sentence-transformers
pyyaml
scikit-learn
scikit-optimize
mysql-connector-python
python-dotenv
qdrant-client
datasets
pytest
python-dotenv